% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{create_sce_from_components}
\alias{create_sce_from_components}
\title{Convert components list to SingleCellExperiment object}
\usage{
create_sce_from_components(components)
}
\arguments{
\item{components}{Named list with X, obs, var, obsm, obsp, etc.}
}
\value{
SingleCellExperiment object
}
\description{
Convert components list to SingleCellExperiment object
}
