% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{create_seurat_from_components}
\alias{create_seurat_from_components}
\title{Convert components list to Seurat object}
\usage{
create_seurat_from_components(components)
}
\arguments{
\item{components}{Named list with X, obs, var, obsm, obsp, etc.}
}
\value{
Seurat object
}
\description{
Convert components list to Seurat object
}
