% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{extract_components_from_sce}
\alias{extract_components_from_sce}
\title{Extract components from SingleCellExperiment object}
\usage{
extract_components_from_sce(sce_obj)
}
\arguments{
\item{sce_obj}{SingleCellExperiment object}
}
\value{
Named list with all components
}
\description{
Extract components from SingleCellExperiment object
}
