% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{extract_components_from_seurat}
\alias{extract_components_from_seurat}
\title{Extract components from Seurat object}
\usage{
extract_components_from_seurat(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat object}
}
\value{
Named list with all components
}
\description{
Extract components from Seurat object
}
