% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{load_json}
\alias{load_json}
\title{Load list from JSON format}
\usage{
load_json(json_file)
}
\arguments{
\item{json_file}{Path to JSON file}
}
\value{
Named list
}
\description{
Load list from JSON format
}
