% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{load_mtx}
\alias{load_mtx}
\title{Load sparse matrix from MTX format}
\usage{
load_mtx(mtx_file)
}
\arguments{
\item{mtx_file}{Path to MTX file (can be gzipped)}
}
\value{
Sparse matrix
}
\description{
Load sparse matrix from MTX format
}
