% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{load_parquet}
\alias{load_parquet}
\title{Load DataFrame from Parquet format}
\usage{
load_parquet(parquet_file)
}
\arguments{
\item{parquet_file}{Path to Parquet file}
}
\value{
Data frame
}
\description{
Load DataFrame from Parquet format
}
