% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_components}
\alias{read_components}
\title{Read components list from .scb file (advanced usage)}
\usage{
read_components(path)
}
\arguments{
\item{path}{Path to .scb file}
}
\value{
Named list with all components
}
\description{
Read components list from .scb file (advanced usage)
}
