% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{save_json}
\alias{save_json}
\title{Save list to JSON format}
\usage{
save_json(data, file_path, pretty = TRUE)
}
\arguments{
\item{data}{Named list to save}

\item{file_path}{Output file path}

\item{pretty}{Whether to pretty-print JSON}
}
\description{
Save list to JSON format
}
