% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{save_mtx}
\alias{save_mtx}
\title{Save sparse matrix to MTX format}
\usage{
save_mtx(matrix, file_path, compress = TRUE)
}
\arguments{
\item{matrix}{Sparse matrix to save}

\item{file_path}{Output file path}

\item{compress}{Whether to gzip compress}
}
\description{
Save sparse matrix to MTX format
}
