% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{save_parquet}
\alias{save_parquet}
\title{Save DataFrame to Parquet format}
\usage{
save_parquet(df, file_path)
}
\arguments{
\item{df}{Data frame to save}

\item{file_path}{Output file path}
}
\description{
Save DataFrame to Parquet format
}
