% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{scio_read_components}
\alias{scio_read_components}
\title{Read components list from .scio file (advanced usage)}
\usage{
scio_read_components(path)
}
\arguments{
\item{path}{Path to .scio file}
}
\value{
Named list with all components
}
\description{
Read components list from .scio file (advanced usage)
}
