% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{scio_write_components}
\alias{scio_write_components}
\title{Write components list to .scio file (advanced usage)}
\usage{
scio_write_components(components, path, overwrite = FALSE, compress = TRUE)
}
\arguments{
\item{components}{Named list with all components}

\item{path}{Output .scio file path}

\item{overwrite}{Whether to overwrite existing file}

\item{compress}{Whether to compress MTX files}
}
\description{
Write components list to .scio file (advanced usage)
}
